pyproject.toml
[project]
name = "time-series-midterm"
version = "0.1.0"
description = ""
authors = [
    {name = "π",email = "pirayan20@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<3.12"
dependencies = [
    "nltk (>=3.9.1,<4.0.0)",
    "matplotlib (>=3.10.0,<4.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "openai (>=1.64.0,<2.0.0)",
    "streamlit (>=1.42.2,<2.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
